+++ vns/service.go
@@ -15,7 +15,7 @@ type Stop struct {
 // TravelCost holds the main properties to describes a cost between two points.
 type TravelCost struct {
 	// Distance represents the distance between two stops in meters.
-	Distance float32
+	Distance float64
 }
 
 // CostMatrix describes a distance matrix between all stops.
